* Source/NSCachedImageRep.m: Restore the implementation of -draw because libart/xlib need it.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@33797 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSCachedImageRep.h
@@ -52,7 +52,7 @@
  */
 - (id) initWithSize: (NSSize)aSize
 	 pixelsWide: (int)pixelsWide
-	 pixelsHigh: (NSInteger)pixelsHigh
+	 pixelsHigh: (int)pixelsHigh
 	      depth: (int)aDepth
 	   separate: (BOOL)separate
 	      alpha: (BOOL)alpha;
